+++ crates/adapters/risc0/src/metrics.rs
@@ -16,6 +16,7 @@ pub fn metrics_callback(input: Bytes) -> anyhow::Result<Bytes> {
         count: cycles_count,
         free_heap_bytes,
     } = sov_metrics::cycle_utils::deserialize_metrics_call(input.as_ref())?;
+
     sov_metrics::track_metrics(|tracker| {
         tracker.track_zkvm_metric(sov_metrics::ZkVmExecutionChunk {
             name: metric,