+++ crates/module-system/sov-modules-macros/tests/integration/metrics.rs
@@ -30,6 +30,12 @@ async fn test_metrics_macro() {
         .await
         .expect("Impossible to bind to port");
 
+    // Insert metadata for the metrics
+    sov_metrics::METRICS_METADATA
+        .write()
+        .unwrap()
+        .insert("test_metadata".to_string(), "test_value".to_string());
+
     init_metrics_tracker(&MonitoringConfig {
         telegraf_address: channel.local_addr().unwrap(),
         max_datagram_size: Some(1),
@@ -56,7 +62,7 @@ async fn test_metrics_macro() {
     let mut parsed_buf = std::str::from_utf8(&buf[..]).unwrap().split(" ");
     assert_eq!(
         parsed_buf.next().unwrap(),
-        "sov_rollup_gas_constant,name=test_metrics,input=\"10\",constant=test"
+        "sov_rollup_gas_constant,name=test_metrics,constant=test,input=10,test_metadata=test_value"
     );
     assert_eq!(parsed_buf.next().unwrap(), "num_invocations=1");
 
@@ -82,7 +88,7 @@ async fn test_metrics_macro() {
     let mut parsed_buf = std::str::from_utf8(&buf[..]).unwrap().split(" ");
     assert_eq!(
         parsed_buf.next().unwrap(),
-        "sov_rollup_gas_constant,name=test_metrics,input=\"100\",constant=test"
+        "sov_rollup_gas_constant,name=test_metrics,constant=test,input=100,test_metadata=test_value"
     );
     assert_eq!(parsed_buf.next().unwrap(), "num_invocations=1");